move tests from CDBStructure to MetadataAndVersioning
